add new mvc controllers and html pages

+++ hw08/src/main/java/ru/otus/hw/services/AuthorServiceImpl.java
@@ -4,17 +4,25 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.otus.hw.models.Author;
 import ru.otus.hw.repositories.AuthorRepository;
+import ru.otus.hw.repositories.BookRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 public class AuthorServiceImpl implements AuthorService {
     private final AuthorRepository authorRepository;
+    private final BookRepository bookRepository;
 
     @Override
     public List<Author> findAll() {
         return authorRepository.findAll();
+    }
+
+    @Override
+    public Optional<Author> findById(String id) {
+        return authorRepository.findById(id);
     }
 
     @Override
@@ -24,7 +32,15 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
+    public Author update(String id, String name) {
+        return authorRepository.save(new Author(id, name));
+    }
+
+    @Override
     public void deleteById(String id) {
+        if (bookRepository.existsByAuthor_Id(id)) {
+            throw new IllegalStateException("Нельзя удалить автора: у него есть книги");
+        }
         authorRepository.deleteById(id);
     }
 }
